Move Digital Pack actions into their own controller
