+++ test/controllers/SubscriptionsTest.scala
@@ -1,21 +1,19 @@
 package controllers
 
 import actions.CustomActionBuilders
-import config.StringsConfig
-import org.scalatest.mockito.MockitoSugar.mock
-import play.api.test.Helpers.{contentAsString, status, stubControllerComponents}
-import scala.concurrent.ExecutionContext.Implicits.global
-import cats.implicits._
-import org.scalatest.{MustMatchers, WordSpec}
-import org.scalatest.OptionValues._
-import org.mockito.Mockito.when
-import play.api.test.FakeRequest
-import play.api.test.Helpers._
-import play.api.mvc.Result
-import services.HttpIdentityService
-import fixtures.TestCSRFComponents
 import admin.SwitchState.On
 import admin.{PaymentMethodsSwitch, Settings, SettingsProvider, Switches}
+import cats.implicits._
+import config.StringsConfig
+import fixtures.TestCSRFComponents
+import org.mockito.Mockito.when
+import org.scalatest.OptionValues._
+import org.scalatest.mockito.MockitoSugar.mock
+import org.scalatest.{MustMatchers, WordSpec}
+import play.api.mvc.Result
+import play.api.test.FakeRequest
+import play.api.test.Helpers.{contentAsString, status, stubControllerComponents, _}
+import services.HttpIdentityService
 
 import scala.concurrent.Future
 
@@ -25,17 +23,17 @@ class SubscriptionsTest extends WordSpec with MustMatchers with TestCSRFComponen
 
     val checkoutEndpoint = "subscribe/digital/checkout?displayCheckout=true"
 
-    def fakeSubscriptions(
+    def fakeDigitalPack(
       actionRefiner: CustomActionBuilders = loggedInActionRefiner,
       identityService: HttpIdentityService = mockedIdentityService(authenticatedIdUser.user -> idUser.asRight[String])
-    ): Subscriptions = {
+    ): DigitalPack = {
 
       val settingsProvider = mock[SettingsProvider]
       when(settingsProvider.settings()).thenReturn(
         Settings(Switches(PaymentMethodsSwitch(On, On, None), PaymentMethodsSwitch(On, On, Some(On)), Map.empty, On))
       )
 
-      new Subscriptions(
+      new DigitalPack(
         actionRefiner,
         identityService,
         assetResolver,
@@ -50,7 +48,7 @@ class SubscriptionsTest extends WordSpec with MustMatchers with TestCSRFComponen
       actionRefiner: CustomActionBuilders = loggedInActionRefiner,
       identityService: HttpIdentityService = mockedIdentityService(authenticatedIdUser.user -> idUser.asRight[String])
     ): Future[Result] = {
-      fakeSubscriptions(actionRefiner, identityService).displayForm("UK", true, true)(FakeRequest())
+      fakeDigitalPack(actionRefiner, identityService).displayForm("UK", true, true)(FakeRequest())
     }
   }
